+++ src/inference.py
@@ -1,0 +1,5 @@
+from generators import DCGANGenerator
+
+
+if __name__=="__main__":
+    pass
